Parsing errors no longer erroneously claim to be compiler bugs
PiperOrigin-RevId: 192495926
