+++ angular/lib/src/compiler/parse_util.dart
@@ -1,3 +1,4 @@
+import 'package:angular_ast/angular_ast.dart';
 import 'package:angular_compiler/cli.dart';
 import 'package:source_span/source_span.dart';
 
@@ -15,4 +16,64 @@ abstract class ParseError extends BuildError {
 
   @override
   String toString() => message;
+}
+
+class AstExceptionHandler extends RecoveringExceptionHandler {
+  final String template;
+  final String sourceUrl;
+
+  final parseErrors = <ParseError>[];
+
+  AstExceptionHandler(this.template, this.sourceUrl);
+
+  void handleParseError(ParseError error) {
+    parseErrors.add(error);
+  }
+
+  void handleAll(Iterable<ParseError> errors) {
+    parseErrors.addAll(errors);
+  }
+
+  void maybeReportExceptions() {
+    if (exceptions.isNotEmpty) {
+      // We always throw here, so no need to clear the list.
+      _reportExceptions();
+    }
+    if (parseErrors.isNotEmpty) {
+      // TODO(alorenzen): Once this is no longer used for the legacy parser,
+      // rename to reportParseErrors.
+      _handleParseErrors(parseErrors);
+      // handleParseErrors() may only log warnings and not throw, so we need to
+      // clear the list before the next phase.
+      parseErrors.clear();
+    }
+  }
+
+  void _reportExceptions() {
+    final sourceFile = new SourceFile.fromString(template, url: sourceUrl);
+    final errorString = exceptions
+        .map((exception) => sourceFile
+            .span(exception.offset, exception.offset + exception.length)
+            .message(exception.errorCode.message))
+        .join('\n');
+    throw new BuildError('Template parse errors:\n$errorString');
+  }
+}
+
+void _handleParseErrors(List<ParseError> parseErrors) {
+  final warnings = <ParseError>[];
+  final errors = <ParseError>[];
+  for (final error in parseErrors) {
+    if (error.level == ParseErrorLevel.WARNING) {
+      warnings.add(error);
+    } else if (error.level == ParseErrorLevel.FATAL) {
+      errors.add(error);
+    }
+  }
+  if (warnings.isNotEmpty) {
+    logWarning('Template parse warnings:\n${warnings.join('\n')}');
+  }
+  if (errors.isNotEmpty) {
+    throw new BuildError('Template parse errors:\n${errors.join('\n')}');
+  }
 }
